YouTube woes (see below)
Related issues:

* rg3/youtube-dl#15275
* rg3/youtube-dl#6923
* rg3/youtube-dl#11476
*
https://www.reddit.com/r/DataHoarder/comments/7qpuhn/youtubedl_being_thr
ottled/?sort=new

—-

So yes, thankfully we don’t care about the filesize or 1080P HD VP9
video. While ‘best’ won’t download .webm, and therefore no OPUS, we’re
using ffmpeg anyways so it’s not that much of a difference if at all —
ffmpeg may be more efficient encoding AAC. The audio with ‘best’ is
still AAC, which is comparable at the bitrate. It’s just less efficient
for YouTube and if you use LavaPlayer then it’s much more of a burden.

One can install Axel (apt-get install axel) and then uncomment the last
two lines and change the first back to ‘bestaudio/best’ — it’s much
faster but the speed is still slow and only benefits larger files. So a
2 hour .webm will take 2 minutes to download vs. who knows. But most
smaller videos still take 45-60 seconds which is kind of annoying for
8MB.

Tl;dr — try out the option if you wish, but ‘best’ is really fine for
now.

YouTube may fix this in the future, will update then.