fixing some playlist issues
One of the issues actually had to do with me using bisection. Bisection used the entry's "sort attribute" to find the index of the entry.
The "sort attribute" being the title of the entry. When searching for the position (index) of a specific entry though, it would fail because
the identity (id/hash) of an entry is determined by its url.