saving playlists when changes occur during loading
This addresses an interesting issue which happens when a playlist's entries don't have an entry_id yet.
If such a playlist is loaded and Giesela shut down then the corresponding entry_ids written to redis won't map
to the correct playlist entries as their entry_ids will be generated again (since it wasn't saved).

With this commit the playlists are saved as soon as they're loaded if there was a change within the playlist.