Just for now..(see description -shrug-)
Discord is working (hopefully soon, it is being worked on) on getting the member ID's who own the bot under the new `teams` structure, and lib-dev's are mostly holding off implementing any hard logic until Discord updates. Anyways, this branch is dev anyways (it's what we run though) and it's safe here as no one else should be running this branch in its current state. Not worth setting the `bot_owner` (which you can of course); this allows admins (only me, @siku2 and some bots have that permission) which makes it easier during development. And yes I could have added two lines and a check for our user ID's, but a lot of this will be changing and I'm lazy, this took 30 seconds and works for now (for ref, I had `@commands.is_owner` flags before on the GieeslaRewrite running bot). 🙃